refactor: Refactored and improved handle byte length unit tests
